requirements.txt
asgiref==3.6.0
async-timeout==4.0.3
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-toolmux==0.1.2
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyTelegramBotAPI==4.20.0
python-dotenv==0.21.1
pytz==2022.7.1
redis==5.0.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
